Simple specs for the Player

+++ lib/waiting_on_rails/player.rb
@@ -1,18 +1,21 @@
 module WaitingOnRails
   class Player
+    attr_reader :pid
+
     def initialize
-      @pid = nil
     end
 
     def start
       @pid = spawn("mplayer #{music_path}", :out => '/dev/null', :err => '/dev/null')
+      Process.detach(@pid)
     end
 
     def stop
+      raise "Player was not started." if @pid.nil?
       Process.kill('TERM', @pid)
       Process.wait
       true
-    rescue Errno::ESRCH
+    rescue Errno::ESRCH, Errno::ECHILD
       false
     end
